Try one-shot modifiers
I'm using Callum's user-space implementation:
https://github.com/callum-oakley/qmk_firmware/tree/d63988caaf86217cfddf903c2c32dc042a8ef4da/users/callum

I did this so that I wouldn't be plagued by the issues I've been having
with holding dual-function keys for too short or too long a time.

I also put the MDIA layer on slash rather than apostrophe, that way a
dual-function ctrl can be on slash to match the LH side of the
keyboard.